moved apps to separate shared package
