+++ apps/photos/src/pages/login/index.tsx
@@ -2,7 +2,7 @@ import LoginPage from '@ente/accounts/pages/login';
 import { useRouter } from 'next/router';
 import { AppContext } from 'pages/_app';
 import { useContext } from 'react';
-import { APPS } from '@ente/shared/constants/apps';
+import { APPS } from '@ente/shared/apps/constants';
 
 export default function Login() {
     const appContext = useContext(AppContext);
