+++ packages/shared/apps/types.ts
@@ -1,0 +1,8 @@
+import { NextRouter } from 'next/router';
+
+export interface PageProps {
+    appContext: {
+        showNavBar: (show: boolean) => void;
+    };
+    router: NextRouter;
+}
